various changes in building techdoc.tex: - make table of contents, footnotes etc. clickable hyperlinks - use timestamp of techdoc.tex (instead of build time) as creationdate   in the PDF metadata - don't include build date on first page of the PDF - make clean:   - delete techdoc.out (created by pdftex)   - fix deletion of techdoc.txt (was techdo_r_.txt)
The initial target was to get reproduceable PDF builds (therefore the 
timestamp-related changes), the other things came up during discussing
this patch with David Haller.

The only remaining difference in the PDF from build to build is the /ID
line.  This line can't be controlled in pdflatex and is now filtered 
out by build-compare in the openSUSE build service (bnc#760867).

Credits go to David Haller for writing large parts of this patch
(but he didn't notice the techdo_r_.txt ;-)


Signed-Off-By: Christian Boltz <apparmor@cboltz.de>